Moved reversing_permutation to separate header in detail
