+++ CMakeLists.txt
@@ -54,7 +54,7 @@ find_package(OpenMP)
 #message(STATUS ">>> OpenMP_CXX_FLAGS = ${OpenMP_CXX_FLAGS}")
 
 # Interface libraries sometimes complain if you list sources, see below.
-add_library(Libalgebra INTERFACE)
+add_library(Libalgebra INTERFACE libalgebra/detail/reversing_permutation.h)
 
 
 # For IDEs that use target sources lists to generate project information
